Reverted back to using border spacing
After some testing I’ve decided using border spacing and negative
margins is the best way to negate the spacing around the first and last
columns. Using borders instead of border-spacing makes the col sizing
funky in the last column if no specific widths are set(due to the last
column not being the same box sizing as the previous cols). I feel like
this defeats the purpose of using display:table in the first place so
this will stick to border spacing indefinitely for now.